Refactor: [/post/{id}] 날짜 포매팅 함수 유틸리티로 분리
