+++ src/components/feature/PostDetail/CardModal.jsx
@@ -1,8 +1,9 @@
 import { createPortal } from 'react-dom';
-import styles from './CardModal.module.css';
+import { DEFAULT_PROFILE_URL } from '@/pages/MessageCreate/MessageCreate';
+import { formatDate } from '@/utils/formatDate';
 import Label from './Label';
 import Button from '@/components/common/Button/Button';
-import { DEFAULT_PROFILE_URL } from '@/pages/MessageCreate/MessageCreate';
+import styles from './CardModal.module.css';
 
 // API 폰트 이름을 CSS 폰트 이름으로 변환
 const getFontFamily = (apiFont) => {
@@ -40,14 +41,6 @@ function CardModal({
   font,
   createdAt,
 }) {
-  // createdAt을 디자인과 같은 형식으로 변환
-  function formatDate(dateString) {
-    const date = new Date(dateString);
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    return `${year}.${month}.${day}`;
-  }
   const data = createdAt ? formatDate(createdAt) : '';
 
   if (!isOpen) return null;
